pyproject.toml
[tool.poetry]
name = "photowall"
version = "0.1.0"
description = ""
authors = ["cyanife <cyanife@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.1.0"
numpy = "^1.19.5"

[tool.poetry.dev-dependencies]
isort = "^5.7.0"
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
